--- a/apps/api/src/app/prisma/schema.prisma
+++ b/apps/api/src/app/prisma/schema.prisma
@@ -17,7 +17,8 @@ model link {
   project      String?  @db.Uuid
   customHashId String?  @unique
   params       Json?
-
+  state        State    @default(Draft)
+  expiry       BigInt?
   @@unique([userID, project, url, customHashId])
 }
 
@@ -25,4 +26,10 @@ model template {
   id   String              @id @default(dbgenerated("gen_random_uuid()")) @db.Uuid
   text String?
   name Unsupported("name")
+}
+
+enum State {
+  Draft
+  Live
+  Expire
 }